+++ src/main/java/com/atlauncher/gui/panels/NoCurseForgePacksPanel.java
@@ -1,0 +1,41 @@
+/*
+ * ATLauncher - https://github.com/ATLauncher/ATLauncher
+ * Copyright (C) 2013-2021 ATLauncher
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.atlauncher.gui.panels;
+
+import java.awt.Component;
+
+import javax.swing.BoxLayout;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+
+import org.mini2Dx.gettext.GetText;
+
+// TODO: this, NoCurseModsPanel and NoFTBPacksPanel can be refactored to be better since they're all similar
+@SuppressWarnings("serial")
+public class NoCurseForgePacksPanel extends JPanel {
+    public NoCurseForgePacksPanel(boolean hasSearch) {
+        setLayout(new BoxLayout(this, BoxLayout.Y_AXIS));
+
+        JLabel label = new JLabel(hasSearch ? GetText.tr("No packs found. Remove your search query and try again.")
+                : GetText.tr("No packs found."));
+        label.setAlignmentX(Component.CENTER_ALIGNMENT);
+        label.setAlignmentY(Component.CENTER_ALIGNMENT);
+
+        add(label);
+    }
+}