+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add setting to hide pack name and version from Instance card
 - Add support for modpacks.ch (FTB) packs [#392]
 - Add suport for updating CurseForge packs [#393]
+- Add in pack browser for CurseForge packs [#394]
 
 ### Fixes
 - Fix exported instances including disabled mods [#420]